pyproject.toml
[project]
name = "savannah-orders"
version = "0.1.0"
description = "Savannah Orders is a backend service designed to manage customer and order data efficiently. The system allows users to create and retrieve customer and order records through a REST API built with FastAPI for accessment purposes."
authors = [{ name = "kenato254", email = "kenedygitonga@gmail.com" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
  "fastapi (>=0.115.7,<0.116.0)",
  "pyjwt[crypto] (>=2.10.1,<3.0.0)",
  "sqlalchemy (>=2.0.37,<3.0.0)",
  "uvicorn (>=0.34.0,<0.35.0)",
  "loguru (>=0.7.3,<0.8.0)",
  "africastalking (>=1.2.9,<2.0.0)",
  "python-dotenv (>=1.0.1,<2.0.0)",
  "cryptography (>=44.0.0,<45.0.0)",
  "python-keycloak (>=5.3.1,<6.0.0)",
  "python-multipart (>=0.0.20,<0.0.21)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{ include = "src" }]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
black = "^25.1.0"
mypy = "^1.14.1"
bandit = "^1.8.2"
pytest-cov = "^6.0.0"
pytest-asyncio = "^0.25.3"
aiosqlite = "^0.20.0"
isort = "^6.0.0"
flake8 = "^7.1.1"

[tool.black]
line-length = 79
include = '\.pyi?$'
skip-string-normalization = true
skip-magic-trailing-comma = true
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/'''


[tool.flake8]
max-line-length = 79


[tool.poetry.scripts]
start = 'src.app.main:main'
test = 'src.tests.run_tests:main'
